meu_sistema_manutencao/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-login==0.6.2
flask-wtf==1.1.1
werkzeug==2.3.4
python-dotenv==1.0.0
email-validator==2.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
pytest==7.3.1
coverage==7.2.5
pytest-cov==4.0.0
flask-testing==0.8.1
flask-swagger-ui==3.36.0
pyyaml==6.0
flasgger==0.9.5
psutil==5.9.5
enum34==1.1.10
celery==5.2.7
redis==4.5.4
